Fixing failure to propagate exceptions across the trampoline.
The issue observed is that the program aborts because the C++ runtime
function __cxa_throw cannot find an exception handler. Runtime errors
in the Zeek scripting language are propagated by exceptions, notably
zeek::InterpreterException. When compiling with exception support, a C++
compiler also generates the contents for the .eh_frame section. This
section supports the exception mechanism by permitting the discovery
of the appropriate catch statement, among other things.

In theory, it should be possible to dynamically register the data necessary
to unwind the stack over the trampoline, but doing so in a robust and portable
manner is not trivial. I believe this would require calling functions
from the C++ runtime library. Furthermore, the data that would need to be
registered would depend on the processor architecture. Using a library that
provides support for JIT compilation and loading might help solve this problem
for the general case, but it doesn't seem to be worth using a library for the
particular case here.

The approach adopted by this commit is that InterpreterException is
caught and the rethrown to avoid unwinding across the trampoline.
This is not an universal solution in the sense that only known exception
classes are handled correctly. Here, only InterpeterException is handled.

The changes in this commit requires that zeek::SuspededException is a
friend class to zeek::InterpreterException, which is a small change in
upstream Zeek. In addition the SuspededException class could be moved to
Zeek, because declaring a friend that is never defined is weird.

I tested this with gcc_15.0.50.20240403-git and Zeek somewhere near version 7.2.2.
I don't have any way of testing the AARCH version.